Add hpa conditions
