+++ collectors/hpa_test.go
@@ -19,7 +19,7 @@ package collectors
 import (
 	"testing"
 
-	autoscaling "k8s.io/api/autoscaling/v1"
+	autoscaling "k8s.io/api/autoscaling/v2beta1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
